backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
sqlalchemy>=2.0,<3
psycopg[binary]>=3.1,<4
alembic>=1.14,<2
passlib[bcrypt]>=1.7,<2
bcrypt==4.0.1
python-jose[cryptography]>=3.3,<4
pytest>=8.0,<9
httpx>=0.27,<1
openpyxl>=3.1,<4
